Fixed the bug of upload avatar
