+++ app/Http/Controllers/Api/UserController.php
@@ -2,20 +2,25 @@
 
 namespace App\Http\Controllers\Api;
 
+use Image;
 use Illuminate\Http\Request;
 use App\Http\Requests\UserRequest;
 use App\Repositories\UserRepository;
 use App\Transformers\UserTransformer;
+use App\Services\FileManager\UploadManager;
 
 class UserController extends ApiController
 {
     protected $user;
 
-    public function __construct(UserRepository $user)
+    protected $manager;
+
+    public function __construct(UserRepository $user, UploadManager $manager)
     {
         parent::__construct();
 
         $this->user = $user;
+        $this->manager = $manager;
     }
 
     /**
@@ -105,5 +110,53 @@ class UserController extends ApiController
         $this->user->destroy($id);
 
         return $this->noContent();
+    }
+
+    /**
+     * Upload the avatar.
+     *
+     * @param Request $request
+     * @return mixed
+     */
+    public function avatar(Request $request)
+    {
+        $file = $request->file('file');
+
+        $validator = \Validator::make([ 'file' => $file ], [ 'file' => 'image' ]);
+
+        if($validator->fails()) {
+            return response()->json([
+                    'success' => false,
+                    'errors'  => $validator->getMessageBag()->toArray()
+                ]);
+        }
+
+        $path = 'avatars/' . auth()->user()->id;
+
+        $result = $this->manager->store($file, $path);
+
+        return $this->respondWithArray($result);
+    }
+
+    /**
+     * Crop Avatar
+     * 
+     * @param  Request $request
+     * @return array
+     */
+    public function cropAvatar(Request $request)
+    {
+        $currentImage = $request->get('image');
+        $data = $request->get('data');
+
+        $image = Image::make($currentImage['relative_url']);
+
+        $image->crop((int) $data['width'], (int) $data['height'], (int) $data['x'], (int) $data['y']);
+
+        $image->save($currentImage['relative_url']);
+
+        $this->user->saveAvatar(auth()->user()->id, $currentImage['url']);
+
+        return $this->respondWithArray($currentImage);
     }
 }
